// appIonic/src/app/models/people.ts
// Modelo People (Taller 3, Parte 1, p.6-7 del PDF)
// Comando origen: ionic g class models/people
// Tipos inferidos de la documentación de SWAPI (https://www.swapi.tech/documentation)
export class People {
  name: string = '';
  height: number = 0;
  mass: number = 0;
  hair_color: string = '';
  skin_color: string = '';
  eye_color: string = '';
  birth_year: string = '';
  gender: string = '';
}