// appIonic/src/app/models/planet.ts
// Modelo Planet (Taller 3, Parte 1, p.6 del PDF)
// Comando origen: ionic g class models/planet
export class Planet {
  climate: string = '';
  diameter: string = '';
  gravity: string = '';
  name: string = '';
  orbital_period: string = '';
  population: string = '';
  rotation_period: string = '';
  surface_water: string = '';
  terrain: string = '';
}