// appIonic/src/app/models/species.ts
// Modelo Species (Taller 3, Parte 1, p.6 del PDF)
// Comando origen: ionic g class models/species
export class Species {
  average_height: string = '';
  average_lifespan: string = '';
  classification: string = '';
  designation: string = '';
  eye_colors: string = '';
  hair_colors: string = '';
  language: string = '';
  name: string = '';
  skin_colors: string = '';
}