// appIonic/src/app/models/starship.ts
// Modelo Starship (Taller 3, Parte 1, p.6 del PDF)
// Comando origen: ionic g class models/starship
export class Starship {
  cargo_capacity: string = '';
  cost_in_credits: string = '';
  crew: string = '';
  hyperdrive_rating: string = '';
  length: string = '';
  manufacturer: string = '';
  model: string = '';
  name: string = '';
  passengers: string = '';
  starship_class: string = '';
}